add save old log files

+++ WebApplication1/FileManager.aspx.cs
@@ -25,7 +25,7 @@ namespace WebApplication1
                 FileInfo[] Files = dinfo.GetFiles(FileType);
                 foreach (FileInfo file in Files)
                 {
-                    if (file.Name != "EMPAC_tracker_config.txt") // dont want to allow deletion of config file.  
+                  //  if (file.Name != "EMPAC_tracker_config.txt") // dont want to allow deletion of config file.  
                         lsb.Items.Add(file.Name);
                 }
             }
@@ -38,8 +38,8 @@ namespace WebApplication1
 
             foreach(ListItem item in ListBox1.Items)
             {
-                if (item.Selected)
-                    File.Delete(Server.MapPath("~/App_Data/") + item.ToString());
+                if (item.Selected && item.ToString() != "EMPAC_tracker_config.txt")
+                        File.Delete(Server.MapPath("~/App_Data/") + item.ToString());
 
             }
             ListBox1.DataBind();
@@ -51,5 +51,46 @@ namespace WebApplication1
 
 
         }
+
+        protected void Button2_Click(object sender, EventArgs e)
+        {
+
+            foreach (ListItem item in ListBox1.Items)
+            {
+                if (item.Selected)
+                //   File.Delete(Server.MapPath("~/App_Data/") + item.ToString());
+                {
+
+
+                    string fileName = Server.MapPath("~/App_Data/") + item.ToString();
+
+                    System.Web.HttpResponse response = System.Web.HttpContext.Current.Response;
+                    response.ClearContent();
+                    response.Clear();
+                    response.ContentType = "text/plain";
+                    response.AddHeader("Content-Disposition", "attachment; filename=" + item.ToString() + ";");
+                    response.TransmitFile(Server.MapPath("~/App_Data/") + item.ToString());
+                    response.Flush();
+                    response.End();
+
+                }
+              
+            }
+
+
+            //Response.Clear();
+            //Response.ClearHeaders();
+
+            //Response.AddHeader("Content-Length", text.Length.ToString());
+            //Response.ContentType = "text/plain";
+
+            ////   Response.AppendHeader("content-disposition", "attachment;filename=\"EMPAC-Tracker Log.txt\"");
+            //Response.AppendHeader("content-disposition", "attachment;filename=\"EMPAC_Tracker" + DateTime.Now.ToString("yyyyMMddHHmm") + ".txt");
+            //Response.Write(text);
+            //File.WriteAllText(Server.MapPath("~/App_Data/" + "EMPAC_Tracker" + DateTime.Now.ToString("yyyyMMddHHmm") + ".txt"), text);
+            //Response.End();
+        }
+
+    
     }
 }
